Refactored directory and files structure of Platform telemetry files
[PL-20720]: Added ownership tags to test cases
